+++ TranslateCS2.Mod/Services/Exports/IExportServiceStrategy.cs
@@ -1,0 +1,10 @@
+using Game.UI.Widgets;
+
+namespace TranslateCS2.Mod.Services.Exports;
+internal interface IExportServiceStrategy {
+    DropdownItem<string>[] GetExportDropDownItems();
+    DropdownItem<string>[] GetExportTypeDropDownItems();
+    void Export(string localeId,
+                string type,
+                string directory);
+}
